Fixed status bar overlay in iOS

+++ platforms/ios/www/js/FormManager.js
@@ -7,22 +7,25 @@ var FormManager = {
 
   removeStoredForm: function()
   {
-    // Remove images from the device
-    var questions = QuestionManager.getQuestions(this.formInProgress.checklistId);
-    var images, storedValue;
-    for (var i = 0; i < questions.length; i++) {
-      storedValue = QuestionManager.getQuestionStoredValue(questions[i].id);
-      if (!Helper.isEmpty(storedValue) && !Helper.isEmpty(storedValue.images)) {
-        images = storedValue.images;
-        for (var j = 0; j < images.length; j++) {
-          FileManager.deleteFile(images[j]);
+    if (this.formInProgress !== null) {
+      
+      // Remove images from the device
+      var questions = QuestionManager.getQuestions(this.formInProgress.checklistId);
+      var images, storedValue;
+      for (var i = 0; i < questions.length; i++) {
+        storedValue = QuestionManager.getQuestionStoredValue(questions[i].id);
+        if (!Helper.isEmpty(storedValue) && !Helper.isEmpty(storedValue.images)) {
+          images = storedValue.images;
+          for (var j = 0; j < images.length; j++) {
+            FileManager.deleteFile(images[j]);
+          }
         }
       }
+      
+      this.formInProgress = null;
+      window.localStorage.removeItem("navalFormInProgress");
+      console.log('cleared navalFormInProgress');
     }
-    
-    this.formInProgress = null;
-    window.localStorage.removeItem("navalFormInProgress");
-    console.log('cleared navalFormInProgress');
   },
 
   getFormInProgressId: function()
